SBERDOMA-795 Common Wrapper added
