+++ apps/condo/lang/en.json
@@ -402,12 +402,12 @@
   "Print": "Print",
   "Excel": "Upload to Excel",
   "Problem": "Problem",
-  "property.Break": "Break",
-  "property.Continue": "Proceed",
-  "property.Import": "Import",
-  "property.ImportError": "An error occurred while processing the document, check the correctness of the data provided or contact the support service",
-  "property.ImportSuccess": "Homes added",
-  "property.Processing": "Data is being processed, do not close the browser tab or turn off the computer",
+  "Break": "Break",
+  "Continue": "Proceed",
+  "Import": "Import",
+  "ImportError": "An error occurred while processing the document, check the correctness of the data provided or contact the support service",
+  "ImportSuccess": "{objects} added",
+  "ImportProcessing": "Data is being processed, do not close the browser tab or turn off the computer",
   "ReadMore": "Expand",
   "Redirecting": "Redirecting ...",
   "Register": "Register",
