+++ apps/condo/lang/ru.json
@@ -402,12 +402,12 @@
   "Print": "Распечатать",
   "Excel": "Выгрузить в Excel",
   "Problem": "Проблема",
-  "property.Break": "Прервать",
-  "property.Continue": "Продолжить",
-  "property.Import": "Импорт",
-  "property.ImportError": "Во время обработки документа произошла ошибка, проверьте корректность предоставленных данных или обратитесь в службу поддержки",
-  "property.ImportSuccess": "Дома добавлены.",
-  "property.Processing": "Данные обрабатываются, не закрывайте вкладку браузера и не выключайте компьютер",
+  "Break": "Прервать",
+  "Continue": "Продолжить",
+  "Import": "Импорт",
+  "ImportError": "Во время обработки документа произошла ошибка, проверьте корректность предоставленных данных или обратитесь в службу поддержки",
+  "ImportSuccess": "{objects} добавлены.",
+  "ImportProcessing": "Данные обрабатываются, не закрывайте вкладку браузера и не выключайте компьютер",
   "ReadMore": "Раскрыть",
   "Redirecting": "Перенаправляем ...",
   "Register": "Зарегистрироваться",
